modified navigation structure
