+++ src/App.js
@@ -2,11 +2,10 @@ import React from 'react';
 import {createAppContainer} from 'react-navigation';
 import {Provider as ReduxProvider} from 'react-redux';
 import configureStore from 'store/configureStore';
-import MainNavigator from './navigation/MainNavigator';
+import AppNavigator from './navigation';
 
 const store = configureStore();
-
-const AppContainer = createAppContainer(MainNavigator);
+const AppContainer = createAppContainer(AppNavigator);
 
 export default function App() {
   return (
