--- src/navigation/HomeNavigator.js
@@ -1,16 +1,0 @@
-import {createDrawerNavigator} from 'react-navigation-drawer';
-
-import HomeScreen from '../screens/home/HomeScreen';
-import TasksScreen from '../screens/tasks/TasksScreen';
-import colors from 'assets/colors';
-
-const HomeNavigator = createDrawerNavigator({
-  HomeScreen,
-  TasksScreen,
-});
-
-HomeNavigator.navigationOptions = {
-  headerShown: false,
-};
-
-export default HomeNavigator;
